refactor: remove UserInfoController.java and move api to AuthController
