+++ src/main/java/com/project/react/controller/AuthController.java
@@ -6,12 +6,16 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.security.core.GrantedAuthority;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -20,6 +24,9 @@ import com.project.react.repository.UserDb;
 import com.project.react.restModel.BaseResponse;
 import com.project.react.security.jwtAuth.AuthenticationRequest;
 import com.project.react.security.jwtAuth.JwtTokenProvider;
+
+import static org.springframework.http.ResponseEntity.ok;
+import static java.util.stream.Collectors.toList;
 
 @RestController
 @CrossOrigin(origins = "http://localhost:3000")
@@ -30,7 +37,7 @@ public class AuthController {
 
     @Autowired
     JwtTokenProvider jwtTokenProvider;
-    
+
     @Autowired
     UserDb users;
 
@@ -49,5 +56,14 @@ public class AuthController {
         } catch (AuthenticationException e) {
             throw new BadCredentialsException("Invalid username/password supplied");
         }
+    }
+
+    @GetMapping("/check")
+    public ResponseEntity<?> currentUser(@AuthenticationPrincipal UserDetails userDetails) {
+        Map<Object, Object> model = new HashMap<>();
+        model.put("username", userDetails.getUsername());
+        model.put("roles", userDetails.getAuthorities().stream().map(a -> ((GrantedAuthority) a).getAuthority())
+                .collect(toList()));
+        return ok(model);
     }
 }
