--- src/main/java/com/project/react/controller/UserInfoController.java
@@ -1,26 +1,0 @@
-package com.project.react.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import static java.util.stream.Collectors.toList;
-import static org.springframework.http.ResponseEntity.ok;
-
-@RestController()
-public class UserInfoController {
-    @GetMapping("/api/me")
-    public ResponseEntity<?> currentUser(@AuthenticationPrincipal UserDetails userDetails) {
-        Map<Object, Object> model = new HashMap<>();
-        model.put("username", userDetails.getUsername());
-        model.put("roles", userDetails.getAuthorities().stream().map(a -> ((GrantedAuthority) a).getAuthority())
-                .collect(toList()));
-        return ok(model);
-    }
-}
